BAU Rename class files to adhere to conventions
Rename these class files to match our existing convention of
`UpperCamelCase.class.js`.
